pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "liquid-crystal-simulator"
version = "0.1.0"
description = "2D simulation framework for nematic liquid crystals using Landau-de Gennes theory"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Ehsan Es'haghi"},
]

dependencies = [
    "numpy>=1.20.0",
    "pandas>=1.3.0",
    "matplotlib>=3.5.0",
    "Pillow>=9.0.0",
]

[project.optional-dependencies]
jax = [
    "jax>=0.4.0",
    "jaxlib>=0.4.0",
    "diffrax>=0.4.0",
]

[project.scripts]
lcs-run = "scripts.run_simulations:main"
lcs-visualize = "scripts.visualize:main"
lcs-dt-finder = "scripts.dt_euler_finder:main"

[tool.setuptools.packages.find]
include = ["src*", "scripts*"]
